# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RT Management API",
        "description": "API documentation for RT Management System",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/laporan/summary": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Get summary pemasukan, pengeluaran, dan saldo per bulan/tahun",
                "operationId": "816ccbe3955e8c55567b2e18c092da6b",
                "parameters": [
                    {
                        "name": "bulan",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tahun",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/laporan/grafik": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Get grafik pemasukan dan pengeluaran per bulan dalam 1 tahun",
                "operationId": "29c2962911b13db919303e3e6e846a42",
                "parameters": [
                    {
                        "name": "tahun",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/laporan/pemasukan": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Get tabel pemasukan per bulan/tahun",
                "operationId": "d0abdb19301571e9a52f0e4a4a8dc332",
                "parameters": [
                    {
                        "name": "bulan",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tahun",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/laporan/pengeluaran": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Get tabel pengeluaran per bulan/tahun",
                "operationId": "fc845a291890db399f61d18961f2f24e",
                "parameters": [
                    {
                        "name": "bulan",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tahun",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/laporan/saldo": {
            "get": {
                "tags": [
                    "Laporan"
                ],
                "summary": "Get saldo berjalan per bulan dalam 1 tahun",
                "operationId": "8cafe2418ede7916e1b1ad284a63750f",
                "parameters": [
                    {
                        "name": "tahun",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/api/pembayaran": {
            "get": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Get list of pembayaran",
                "operationId": "ced6d400ae6007a7789463add1f92e91",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Create new pembayaran",
                "operationId": "99dbd91a19e6f393f2bfc266f7b14805",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rumah_id",
                                    "penghuni_id",
                                    "bulan",
                                    "tahun",
                                    "jenis_iuran",
                                    "jumlah",
                                    "status"
                                ],
                                "properties": {
                                    "rumah_id": {
                                        "type": "integer"
                                    },
                                    "penghuni_id": {
                                        "type": "integer"
                                    },
                                    "bulan": {
                                        "type": "integer"
                                    },
                                    "tahun": {
                                        "type": "integer"
                                    },
                                    "jenis_iuran": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "tanggal_bayar": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    }
                }
            }
        },
        "/api/pembayaran/{id}": {
            "get": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Get detail pembayaran",
                "operationId": "2b3ff503f0051a2f3ded259e5d54acaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Update pembayaran",
                "operationId": "9703cd05d4bee4ad88ccecfdc6af7dcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "rumah_id": {
                                        "type": "integer"
                                    },
                                    "penghuni_id": {
                                        "type": "integer"
                                    },
                                    "bulan": {
                                        "type": "integer"
                                    },
                                    "tahun": {
                                        "type": "integer"
                                    },
                                    "jenis_iuran": {
                                        "type": "string"
                                    },
                                    "jumlah": {
                                        "type": "number"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "tanggal_bayar": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pembayaran"
                ],
                "summary": "Delete pembayaran",
                "operationId": "f9e73cdb4e374f77fb2223c7c62cc111",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    }
                }
            }
        },
...
    },
...
}